requirements.txt
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
jsonpickle
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity
